requirements.txt
django-debug-toolbar==1.11.1
asgiref
autopep8
django==3.2.5
django-admin-sso==3.1.0
adminlte-base==0.2.0
arrow==1.2.1
django-adminlte-full==0.2.0
django-crispy-forms==1.13.0
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dateutil==2.8.2
pycodestyle
pytz
sqlparse
Unipath~=1.1
dj-database-url
python-decouple
gunicorn==20.1.0
whitenoise==5.3.0
django-import-export
django-filter~=21.1
djangorestframework~=3.12.4
Markdown==3.3.6
django-tables2==2.4.1
mysqlclient==2.1.0
psycopg2-binary==2.8.5
reportlab==3.6.3
pillow==8.4.0
django-crispy-forms==1.13.0
django-ckeditor==6.2.0
djangoajax==3.2
django_ajax==0.3.0
django-ajax-tables==1.1.1
